requirements.txt
groq
python-dotenv
chainlit
numpy